chore: move billing to new module structure
