+++ apps/web/app/(app)/environments/[environmentId]/settings/(organization)/billing/page.tsx
@@ -1,0 +1,3 @@
+import PricingPage from "@/modules/ee/billing/page";
+
+export default PricingPage;
