--- apps/web/app/(ee)/(billing)/environments/[environmentId]/layout.tsx
@@ -1,3 +1,0 @@
-import EnvLayout from "../../../../(app)/environments/[environmentId]/layout";
-
-export default EnvLayout;
