+++ apps/web/app/api/billing/stripe-webhook/route.ts
@@ -1,6 +1,6 @@
 import { responses } from "@/app/lib/api/response";
+import { webhookHandler } from "@/modules/ee/billing/api/lib/stripeWebhook";
 import { headers } from "next/headers";
-import { webhookHandler } from "./lib/stripeWebhook";
 
 export const POST = async (request: Request) => {
   const body = await request.text();
