+++ apps/web/modules/ee/billing/components/PricingTable.tsx
@@ -1,11 +1,5 @@
 "use client";
 
-import { CLOUD_PRICING_DATA } from "@/app/(ee)/(billing)/api/billing/stripe-webhook/lib/constants";
-import {
-  isSubscriptionCancelledAction,
-  manageSubscriptionAction,
-  upgradePlanAction,
-} from "@/app/(ee)/(billing)/environments/[environmentId]/settings/(organization)/billing/actions";
 import { useTranslations } from "next-intl";
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
@@ -17,6 +11,8 @@ import { Badge } from "@formbricks/ui/components/Badge";
 import { BillingSlider } from "@formbricks/ui/components/BillingSlider";
 import { Button } from "@formbricks/ui/components/Button";
 import { PricingCard } from "@formbricks/ui/components/PricingCard";
+import { isSubscriptionCancelledAction, manageSubscriptionAction, upgradePlanAction } from "../actions";
+import { CLOUD_PRICING_DATA } from "../api/lib/constants";
 
 interface PricingTableProps {
   organization: TOrganization;
